DTSPB-2528: refactor OCRFieldIhtMoneyMapper
